Moved useDisappearingNotification hook to the hooks folder
